Added to convert the old style to the new.
The old style listed the adjacency lists for all reactions,
but the reactants and products are often repeated throughout.
This was changed to use a representation of the reaction as a
string, then source each reactant and product adjacency list
(including atom labels) from a species dictionary. This shrinks
the training data size as it avoids repeated species.

There is still some uncertainty if the same species with
different labeling can be properly distinguished. There might be
a possibility of cases where these are not handled
properly (should be rare, but still an issue).